feat: move storage interface into a separate file-s
