+++ libs/ledger-live-common/src/DataModel.ts
@@ -62,15 +62,3 @@ export function createDataModel<R, M>(
     encode: encodeModel,
   });
 }
-
-export interface SimpleStorage {
-  kind: "simple";
-  save(key: string, value: unknown): Promise<void>;
-  get(key: string): Promise<unknown>;
-}
-
-export interface NamedSpaceStorage {
-  kind: "namedspace";
-  save(ns: string, key: string, value: unknown): Promise<void>;
-  get(ns: string, key: string): Promise<unknown>;
-}
